Nouveau controller pour page d'ajout, non fonctionnel

+++ src/computer/database/domain/Company.java
@@ -14,15 +14,15 @@ public class Company {
 
 	@Id 
 	@GeneratedValue
-	private int id;
+	private long id;
 	
 	@Column(name="name")
 	private String name;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getName() {
@@ -39,7 +39,7 @@ public class Company {
 			company = new Company();
 		}
 		
-		public Builder id(int id) {
+		public Builder id(long id) {
 			company.setId(id);
 			return this;
 		}
